pyproject.toml
[project]
name = "welcomehomedb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "flask>=3.1.0",
    "flask-login>=0.6.3",
    "mysql-connector-python>=9.1.0",
    "pillow>=11.0.0",
    "matplotlib>=3.9.3",
    "ipykernel>=6.29.5",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.1",
]
